security-ai/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Solana and blockchain
solana==0.30.2
solders==0.18.1
anchorpy==0.19.1

# AI and ML
anthropic==0.7.8
openai==1.3.7
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.3
pandas==2.1.4

# Database and caching
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# HTTP and networking
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2
tenacity==8.2.3

# Security and validation
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
PyNaCl==1.5.0
base58==2.1.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1